+++ rpg-app/src/main/java/com/rpg/rpg_app/user/service/UserService.java
@@ -2,8 +2,10 @@ package com.rpg.rpg_app.user.service;
 
 import com.rpg.rpg_app.user.entity.User;
 import com.rpg.rpg_app.user.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.util.List;
 
@@ -13,27 +15,50 @@ public class UserService implements UserIService {
 
     private final UserRepository userRepository;
 
+    //Função para validação de campos.
+    private void validation(User user) {
+        if (user == null ||
+                !StringUtils.hasText(user.getUsername()) ||
+                !StringUtils.hasText(user.getPassword()) ||
+                !StringUtils.hasText(user.getEmail())) {
+            throw new IllegalArgumentException("Os  campos obrigatórios não podem estar vazios.");
+        }
+
+    }
+
     //Função teste para testar se a 'api' funciona.
     public String printUserName(String username) {
+
         return "Bem vindo usuário :" + username;
+
     }
 
     //Função de salvar usuario.
+    @Override
     public void save(User user) {
-
+        validation(user);
         userRepository.save(user);
 
     }
 
     //Função de dar update num usuario.
+    @Override
     public void update(User user) {
+        if (user.getId() == null || !userRepository.existsById(user.getId())) {
+            throw new EntityNotFoundException("O usuario não foi encontrado para atualizar.");
+        }
 
+        validation(user);
         userRepository.save(user);
 
     }
 
     //Funcao de apagar um usuario.
+    @Override
     public void delete(User user) {
+        if (user == null || user.getId() == null) {
+            throw new IllegalArgumentException("O usuário não foi econtrado para deletar.");
+        }
 
         userRepository.delete(user);
 
@@ -41,12 +66,16 @@ public class UserService implements UserIService {
 
     //Função de encontrar usuario por ID.
     public User findById(Long id) {
+        if(!userRepository.existsById(id)) {
+            throw new IllegalArgumentException("O usuário não foi encontrado.");
+        }
 
         return userRepository.findById(id).orElse(null);
 
     }
 
     //Função para retornar todos os usuários.
+    @Override
     public List<User> findAll() {
 
         return userRepository.findAll();
@@ -54,6 +83,7 @@ public class UserService implements UserIService {
     }
 
     //Função para encontrar usuário por username.
+    @Override
     public List<User> findByUsername(String username) {
 
         return userRepository.findByUsername(username);
@@ -61,6 +91,7 @@ public class UserService implements UserIService {
     }
 
     //Função para encontrar usuario por email.
+    @Override
     public List<User> findByEmail(String email) {
 
         return userRepository.findByEmail(email);
